requirements.txt
Flask==2.3.3
pymongo==4.6.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
PyJWT>=2.0.0
flask-cors>=3.0.0; python_version >= "3.7"